requirements.txt
langchain
langgraph
langchain-google-genai
langchain-core
langchain-community
python-dotenv
streamlit
tavily-python
ipykernel